+++ projects/frontend/es-app/src/guis/menus/GuiMenuCRT.h
@@ -16,7 +16,6 @@ template<class T> class OptionListComponent;
 class GuiMenuCRT : public GuiMenuBase
                  , private IOptionListComponent<CrtAdapterType>
                  , private IOptionListComponent<std::string>
-                 , private ISliderComponent
                  , private ISwitchComponent
                  , private IGuiMenuBase
 {
@@ -59,6 +58,7 @@ class GuiMenuCRT : public GuiMenuBase
     //! Force HDMI video
     bool mForceHDMI;
     bool mOriginalForceHDMI;
+    bool mOriginalFrontendIn240pOn31kHz;
 
     //! Get dacs
     static std::vector<ListEntry<CrtAdapterType>> GetDacEntries(bool onlyRgbDual);
@@ -88,12 +88,6 @@ class GuiMenuCRT : public GuiMenuBase
      */
 
     void OptionListComponentChanged(int id, int index, const std::string & value) override;
-
-    /*
-     * IOptionListComponent<int> implementation
-     */
-
-    void SliderMoved(int id, float value) override;
 
     /*
      * ISwitchComponent implementation