+++ projects/frontend/es-core/src/CrtConf.cpp
@@ -46,6 +46,47 @@ const std::string& CrtConf::CrtAdapterFromEnum(CrtAdapterType adapter)
     case CrtAdapterType::None:
     default: break;
   }
-  static std::string sDefault = "";
+  static std::string sDefault;
   return sDefault;
 }
+
+CrtResolution CrtConf::CrtResolutionFromString(const std::string& menu)
+{
+  if (menu == "p1920x224") return CrtResolution::r224p;
+  if (menu == "p1920x240") return CrtResolution::r240p;
+  if (menu == "p320x240") return CrtResolution::r320x240p;
+  if (menu == "p1920x288") return CrtResolution::r288p;
+  if (menu == "p384x288") return CrtResolution::r384x288p;
+  if (menu == "i640x480") return CrtResolution::r480i;
+  if (menu == "i768x576") return CrtResolution::r576i;
+  if (menu == "p640x480") return CrtResolution::r480p;
+  if (menu == "p1920x240at120") return CrtResolution::r240p120Hz;
+  return CrtResolution::rNone;
+}
+
+const std::string& CrtConf::CrtResolutionFromEnum(CrtResolution type)
+{
+  switch(type)
+  {
+    case CrtResolution::r224p: { static std::string result("p1920x224"); return result; }
+    case CrtResolution::r240p: { static std::string result("p1920x240"); return result; }
+    case CrtResolution::r320x240p: { static std::string result("p320x240"); return result; }
+    case CrtResolution::r288p: { static std::string result("p1920x288"); return result; }
+    case CrtResolution::r384x288p: { static std::string result("p384x288"); return result; }
+    case CrtResolution::r480i: { static std::string result("i640x480"); return result; }
+    case CrtResolution::r576i: { static std::string result("i768x576"); return result; }
+    case CrtResolution::r480p: { static std::string result("p640x480"); return result; }
+    case CrtResolution::r240p120Hz: { static std::string result("p1920x240at120"); return result; }
+    case CrtResolution::_rCount:
+    case CrtResolution::rNone:
+    default: break;
+  }
+  static std::string result("None");
+  return result;
+}
+
+DefineCrtModeOffsetImplementation(VerticalOffset, int, Int, sVerticalOffset, 0)
+DefineCrtModeOffsetImplementation(HorizontalOffset, int, Int, sHorizontalOffset, 0)
+DefineCrtViewportImplementation(Width, int, Int, sWidth, 0)
+
+